Fix app webhook being deleted after updating message

+++ services/backend-api/src/features/feed-connections/feed-connections-discord-channels.service.ts
@@ -703,7 +703,10 @@ export class FeedConnectionsDiscordChannelsService {
         );
       }
 
-      if (oldConnection.details.webhook?.isApplicationOwned) {
+      if (
+        createdApplicationWebhookId &&
+        oldConnection.details.webhook?.isApplicationOwned
+      ) {
         try {
           await this.discordWebhooksService.deleteWebhook(
             oldConnection.details.webhook.id
